fix(coding-agent): resolve ambiguous session selectors
fixes #479
